+++ src/components/admin-navigation/index.tsx
@@ -1,21 +1,3 @@
-import React from 'react';
-import { Categories, setActiveIndex, useAppDispatch } from 'src/store';
-
-const AdminNavigation = () => {
-  const dispatch = useAppDispatch();
-  return (
-    <>
-      <button onClick={() => dispatch(setActiveIndex(Categories.Occurrences))}>
-        Gerichte
-      </button>
-      <button onClick={() => dispatch(setActiveIndex(Categories.Pictures))}>
-        Bilder
-      </button>
-      <button onClick={() => dispatch(setActiveIndex(Categories.Comments))}>
-        Kommentare
-      </button>
-    </>
-  );
-};
+import AdminNavigation from './AdminNavigation';
 
 export default AdminNavigation;
