+++ src/components/navigation/language-switcher/LanguageSwitcher.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { useState } from 'react';
 
-import Icon from '../../util/Icon';
+import Icon from '../../icon/';
 
 import styles from './LanguageSwitcher.module.scss';
