+++ src/components/navigation/language-switcher/index.ts
@@ -1,0 +1,3 @@
+import LanguageSwitcher from './LanguageSwitcher';
+
+export default LanguageSwitcher;
